+++ cli/internal/plugin/manageddestination/windows.go
@@ -16,9 +16,9 @@ func getSysProcAttr() *syscall.SysProcAttr {
 func (c *Client) terminateProcess() error {
 	c.logger.Debug().Msg("sending kill signal to destination plugin")
 	if err := c.cmd.Process.Kill(); err != nil {
-		c.logger.Error().Err(err).Msg("failed to kill source plugin")
+		c.logger.Error().Err(err).Msg("failed to kill destination plugin")
 	}
-	c.logger.Info().Msg("waiting for source plugin to terminate")
+	c.logger.Info().Msg("waiting for destination plugin to terminate")
 	st, err := c.cmd.Process.Wait()
 	if err != nil {
 		return err
@@ -26,7 +26,7 @@ func (c *Client) terminateProcess() error {
 	if !st.Success() {
 		// on windows there is no way to shutdown gracefully via signal. Maybe we can do it via grpc api?
 		// though it is a bit strange to expose api to shutdown a server :thinking?:
-		c.logger.Info().Msgf("source plugin process exited with %s", st.String())
+		c.logger.Info().Msgf("destination plugin process exited with %s", st.String())
 	}
 
 	return nil