+++ plugins/destination/clickhouse/client/client.go
@@ -8,9 +8,9 @@ import (
 
 	"github.com/ClickHouse/clickhouse-go/v2"
 	"github.com/ClickHouse/clickhouse-go/v2/lib/proto"
-	"github.com/cloudquery/plugin-sdk/plugins/destination"
-	"github.com/cloudquery/plugin-sdk/schema"
-	"github.com/cloudquery/plugin-sdk/specs"
+	"github.com/cloudquery/plugin-pb-go/specs"
+	"github.com/cloudquery/plugin-sdk/v2/plugins/destination"
+	"github.com/cloudquery/plugin-sdk/v2/schema"
 	"github.com/rs/zerolog"
 )
 
@@ -26,7 +26,7 @@ type Client struct {
 
 var _ destination.Client = (*Client)(nil)
 
-func (*Client) DeleteStale(context.Context, schema.Tables, string, time.Time) error {
+func (*Client) DeleteStale(context.Context, schema.Schemas, string, time.Time) error {
 	return errors.New("DeleteStale is not implemented")
 }