+++ plugins/destination/clickhouse/typeconv/ch/types/definitions.go
@@ -1,0 +1,28 @@
+package types
+
+import (
+	"github.com/apache/arrow/go/v13/arrow"
+	"github.com/cloudquery/cloudquery/plugins/destination/clickhouse/util"
+)
+
+func FieldType(field arrow.Field) (string, error) {
+	dt, err := ColumnType(field.Type)
+	if err != nil {
+		return "", err
+	}
+
+	if canBeNullable(field.Type) && field.Nullable {
+		return "Nullable(" + dt + ")", nil
+	}
+
+	return dt, nil
+}
+
+func FieldDefinition(field arrow.Field) (string, error) {
+	fieldType, err := FieldType(field)
+	if err != nil {
+		return "", err
+	}
+
+	return util.SanitizeID(field.Name) + " " + fieldType, nil
+}