+++ plugins/source/aws/policies/queries/apigateway/api_gw_execution_logging_enabled.sql
@@ -7,7 +7,7 @@ insert into aws_policy_results
     r.account_id,
     'arn:' || 'aws' || ':apigateway:' || r.region || ':/restapis/' || r.id as resource_id,
     case
-        when s.logging_level not in ('ERROR', 'INFO') then 'fail'
+        when s.logging_level not in ('"ERROR"', '"INFO"') then 'fail'
         else 'pass'
     end as status
 from
@@ -26,7 +26,7 @@ union
      a.account_id,
      'arn:' || 'aws' || ':apigateway:' || a.region || ':/apis/' || a.id as resource_id,
      case
-         when s.route_settings->>'LoggingLevel' in (NULL, 'OFF') then 'fail'
+         when s.default_route_settings->>'LoggingLevel' in (NULL, 'OFF') then 'fail'
          else 'pass'
          end as status
 from